requirements.txt
django-grappelli
Django==4.2.4
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.14.0
django-filter==23.2
markdown==3.4.4
pandas==2.1.0
django-group-by==0.3.1



